services/myservices/sensor/requirements.txt
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
Flask==3.0.3
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
#MarkupSafe==3.0.1 #ta a dar merda
paho-mqtt==1.6.1
pip==24.2
requests==2.32.3
urllib3==2.2.3
Werkzeug==3.0.4
prometheus_client== 0.21.0
